+++ src_new/triggers/levi_ataxia/for_levi/leviticus/passive_active/016_Slough_(Fire_Lord).lua
@@ -42,10 +42,8 @@ local name = matches[2]
 if isTargeted(matches[2]) then
 	-- Slough cures weariness + 1 random affliction
 	erAff("weariness")
-	if removeAffV2 then removeAffV2("weariness") end
 	ataxiaTemp.randomCure = 1
-	if reduceRandomAffCertaintyV2 then reduceRandomAffCertaintyV2() end
-	if onPassiveCureV3 then onPassiveCureV3(1) end
+	onClassCureV3({"weariness"}, 1)
 	selectString(line,1)
 	fg("NavajoWhite")
 	resetFormat()