+++ src_new/triggers/levi_ataxia/for_levi/leviticus/passive_active/018_Continuation_(Monk).lua
@@ -43,9 +43,7 @@ local class = (ataxiaNDB_getClass(name) or "Unknown")
 if isTargeted(matches[2]) and class == "Monk" then
   erAff("weariness")
   ataxiaTemp.randomCure = 1
-  if removeAffV2 then removeAffV2("weariness") end
-  if reduceRandomAffCertaintyV2 then reduceRandomAffCertaintyV2() end
-  if onPassiveCureV3 then onPassiveCureV3(1) end
+  onClassCureV3({"weariness"}, 1)
   selectString(line,1)
   fg("NavajoWhite")
   resetFormat()