+++ src_new/triggers/levi_ataxia/for_levi/leviticus/passive_active/020_Sylvan_Root.lua
@@ -44,9 +44,7 @@ if isTargeted(name) and class == "Sylvan" then
   -- Sylvan root cures haemophilia + 1 random
   erAff("haemophilia")
   ataxiaTemp.randomCure = 1
-  if removeAffV2 then removeAffV2("haemophilia") end
-  if reduceRandomAffCertaintyV2 then reduceRandomAffCertaintyV2() end
-  if onPassiveCureV3 then onPassiveCureV3(1) end
+  onClassCureV3({"haemophilia"}, 1)
   selectString(line,1)
   fg("NavajoWhite")
   resetFormat()